Move ChooseAccount to unread widget module
