+++ feature/widget/unread/src/main/AndroidManifest.xml
@@ -4,6 +4,12 @@
     <application>
 
         <activity
+            android:name=".UnreadWidgetChooseAccountActivity"
+            android:configChanges="locale"
+            android:label="@string/choose_account_title"
+            android:noHistory="true" />
+
+        <activity
             android:name=".UnreadWidgetConfigurationActivity"
             android:exported="false">
             <intent-filter>
